backend/requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
psycopg==3.2.3
psycopg-binary==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
